+++ src/models/vote.rs
@@ -8,7 +8,7 @@ use crate::{models::rating::Rating, prelude::*};
 pub struct Vote {
     pub account_id: u32,
 
-    pub tank_id: i32,
+    pub tank_id: u16,
 
     #[serde(with = "serde_helpers::chrono_datetime_as_bson_datetime")]
     pub timestamp: DateTime,
@@ -21,7 +21,7 @@ pub struct Vote {
 }
 
 impl Vote {
-    pub fn new(account_id: u32, tank_id: i32, rating: Rating) -> Self {
+    pub fn new(account_id: u32, tank_id: u16, rating: Rating) -> Self {
         Self {
             account_id,
             tank_id,
