+++ src/wg/stats.rs
@@ -13,7 +13,7 @@ use crate::{prelude::*, wg::Wg};
 /// Partial user's vehicle statistics.
 #[derive(Deserialize)]
 pub struct VehicleStats {
-    pub tank_id: i32,
+    pub tank_id: u16,
     pub last_battle_time: i64,
 
     #[serde(rename = "all")]
@@ -51,7 +51,7 @@ pub struct InnerVehicleStats {
 #[derive(Clone)]
 pub struct VehicleStatsGetter {
     wee_gee: Wg,
-    cache: Cache<u32, Arc<IndexMap<i32, VehicleStats>>>,
+    cache: Cache<u32, Arc<IndexMap<u16, VehicleStats>>>,
 }
 
 impl From<Wg> for VehicleStatsGetter {
@@ -69,7 +69,7 @@ impl From<Wg> for VehicleStatsGetter {
 impl VehicleStatsGetter {
     /// Retrieve the account's vehicle's statistics and cache it.
     #[instrument(skip_all, fields(account_id = account_id))]
-    pub async fn get(&self, account_id: u32) -> Result<Arc<IndexMap<i32, VehicleStats>>> {
+    pub async fn get(&self, account_id: u32) -> Result<Arc<IndexMap<u16, VehicleStats>>> {
         self.cache
             .try_get_with(account_id, async {
                 let map = self
@@ -89,7 +89,7 @@ impl VehicleStatsGetter {
     }
 
     #[instrument(skip_all, fields(account_id = account_id, tank_id = tank_id))]
-    pub async fn owns_vehicle(&self, account_id: u32, tank_id: i32) -> Result<bool> {
+    pub async fn owns_vehicle(&self, account_id: u32, tank_id: u16) -> Result<bool> {
         Ok(self.get(account_id).await?.get(&tank_id).is_some_and(VehicleStats::is_played))
     }
 }
